api_yamdb/requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
